requirements.txt
git+https://github.com/modelcontextprotocol/protocol-python.git#egg=mcp
httpx
python-dotenv 